Add org.apache.bcel for aswb:aswb_api_integrity_test
go/fix-aswb-tests

ASwB integrity tests find plugin problems that would manifest
as LinkageErrors at runtime. They use the org.apache.bcel
library to inspect bytecode. This library is in Google3 and
not in repo causing build failures in the code recently
migrated to repo. This adds the library in repo.

This CL was created following the following instructions:
https://android.googlesource.com/platform/tools/base/+/mirror-goog-studio-main/bazel/README.md#fetching-new-maven-dependencies

Bug: b/335388630
Test: bazel build //tools/vendor/google3/aswb/javatests/com/google/devtools/intellij/blaze/plugin/aswb:aswb_api_integrity_test
Change-Id: I30ca840aea2161fd1dc399d87d9a3a181d453d63