Split VariantAwareTask interface into VariantTask and GlobalTask
Previously, the VariantAwareTask interface was used for not only variant
tasks but also global (non-variant) tasks. The intention was probably to
reuse the analytics service that VariantAwareTask has for global tasks.

This leads to confusing task hierarchy.

To fix this issue, this change will remove VariantAwareTask, and
introduce 2 new interfaces: VariantTask and GlobalTask, to represent the
2 types of tasks respectively.

Bug: 300617088 (In progress)
Test: Updated TaskMethodModifiersAndAnnotationsTest
Change-Id: Ie3988d240dccc5882bc70b999e1111fcde2195b8