Rename PROCESSED_RES to LINKED_RESOURCES_BINARY_FORMAT
to make it clear that it is the output of `aapt2 link` and it is in
binary format (as opposed to proto format).

Similarly, also rename LINKED_RES_FOR_BUNDLE to
LINKED_RESOURCES_FOR_BUNDLE_PROTO_FORMAT to make it clear that it is in
proto format.

The distiction between the two formats is important because R8 resource
shrinker only works with resources in proto format.

Bug: 326033134
Test: Updated
Change-Id: Ia31aa7fefb14400b1e4bd39b6b35d63317bb72ed