[SDK Index] Do not show URL for non SDK Index libraries
The provided link is to prevent older versions of Studio to get an empty
link and thus there is no point in generating a link if it does not
provide information. With this change first we first check if the link
does not come from the SDK Index and and otherwise returns it.

This change is based on commit 13e6fad89257ba96485e93373739560cad12037f
with some differences in GradleDetectorTest to make sure SDK from
different sources are used.

Bug: 339094763
Test: GooglePlaySdkIndexTest, GradleDetectorTest
Change-Id: Ic969663cd6e724cf2c6f0549d05c75bcfe5a6bad