Fix bug with startup_options being overridden
startup_options is made an internal detail of BazelCmd. Start options should always be the same to prevent bazel starting a different server configuration. A cache-silo-key is added for windows to address --output_base issues described in http://g/android-devtools-infra/c/9PgZVzPv_To - Waiting on a bazel fix to be available in a new release.

Change-Id: Ib3e0253d8d48a46f643c3435639aa7d63027d76e