[fused lib] Fix publication configurations
Gradle 8.8 deprecates use of beforeResolve for mutating configurations.
However, the migration caused no dependencies to be added to the unmerged configuration,
this was unnoticed due to lack of testing.
This change adds the required unmerged dependencies lazily to the unmerged configurations to
fix the issue and adds tests e.g verifying pom dependencies to prevent regressions.

Bug: 346984784
Test: FusedLibraryTest
Change-Id: Ib4514d7bdb48f64bdb89b878e8e5b148b6c1f22f