Introduce ProjectGraph model in AGP
Normally, we initially need the dependencies of each app to figure out
which variants to request for each sub-project, however this initial
fetch of dependencies might take a long time (>5-10 seconds) for
builds with many subprojects. In this case, we ask for a lighter weight
dependency graph (i.e. ProjectGraph) that only contains the sub project
info. This allows the IDE to parallelize execution much faster and start
paralleling in ~2 seconds.

Mentioned values are tested on my machine with 2000 subset of the XL
benchmark.

Test: Covered by snapshot tests
Bug: 354906596
Change-Id: Iaec96137ce3f00fefd2d22a1a45422a96efcbc5f