Run integration tests with project isolation enabled(patch 1)
With this CL, integration tests are running with project isolation by
default and we are using a strict mode, which means any config cache or
project isolation issues would cause the test to fail.

This CL renames the previously used PROJECT_ISOLATION to
PROJECT_ISOLATION_WARN to better reflect the fact that it allows for
config cache problems.

As mentioned above, the strict mode, rather than the warn mode, is used
by default and this ensures no config cache issue to be introduced while
we migrate our tests to run with project isolation.

For the next step, we could look through each "disabled" test and fix test
set up issues and use the PROJECT_ISOLATION_WARN with explicitly problem
count for known issues, or even use PROJECT_ISOLATION if we can get rid
of all of them.

Bug: 270586835
Test: existing
Change-Id: I7d511f6a879c9b913f41c20aa091f47a7530fb04