Improve error handling for AVD home folder
AndroidLocationsException is a checked exception thrown by AvdManager's
constructor under two conditions: sdkManager.location (a final field) is
null, or baseAvdFolder is a regular file rather than a directory.

There is no reason to accept a null sdkManager.location; this should be
considered a programming error, since it is fully in the control of the
programmer to prevent. In contrast, the filesystem can change
dynamically; baseAvdFolder can become valid or invalid during the life
of the AvdManager.

Thus, we should throw an unchecked IllegalArgumentException for a null
SDK location.

It also doesn't make sense to prevent construction of the AvdManager for
what could be a temporary condition. We must handle the case of this
error developing post-construction anyway; throwing an exception on
construction multiplies the number of places we need to consider this
error, and may mislead us into thinking that it cannot occur after
construction. Instead, we can continue to throw when an action is
performed that requires the AVD home directory.

(Of course, this is quite an unlikely condition to begin with. The
primary motivation for this change is to simplify AvdManagerConnection.)

Finally, we should also throw in the case where we cannot create the AVD
directory. This is at least as serious an error as the other case, and
far more likely.

Bug: 339875083
Test: n/a
Change-Id: Ic8f00810dff231babac45f8071d69854a0ccc3fa