[fix] Fail if app has dynamic feature resources, but features are not configured
Change addresses two cases:

When an app depends on another app module, resource linking can fail as
it assumes the app is a feature module. However, it is unclear that this
is the case.

When an app depends on a dynamic feature, but dynamic features are not
configured in the app.

Change adds build failure that prevents the task executing and AAPT
failing during linking; making it unclear of the cause.

Bug: 189499971
Test: AppToAppDependenciesTest
Change-Id: I033bb0b58ccb392f8410d9eff9822fc143739abc