Update jar bytecode verifier to handle androidx-nav
This CL cherrypicks 3 CLs back from 8.8+ to handle the bug
where there's a lot of log spam from the bytecode verifier;
this happens because as of androidx.navigation:navigation-compose:2.8.3
some code moved around in that library which needs adjustment
in the bytecode migration code; this backports it to also be
handled in 8.7.0. Without this, in addition to logspam, those
lint checks are all disabled.

The included CLs:
* Ib8460d8f21b5c1ac037a16ef5d4c388aa1388818
  Add additional regression test for bytecode verifier
  (this is just a test; included to avoid merge
  conflicts on next CL)

* I36008922f6fc0eabb7c1030b2aea1951e37094ab
  More accurately determine whether to use invokeinterface

* I2cc6aaecf69411665fa71f6fe1ca2e06c41b4b57
  Handle LintUtilKt in any package under androidx/navigation/

Bug: 374488858
Test: Existing (and included with this CL)
Change-Id: I4f0714d917d4d11888cab50cc0c2961cf5936d52