Lint: update partial analysis and Detector KDoc
Add to KDoc of Detector functions:

 * afterCheckRootProject
 * checkPartialResults
 * checkMergedProject

to clarify when they are called in global, partial, and isolated
analysis modes. Also, update CommunicationDeviceDetector (the reference
example for partial analysis) to override checkMergedProject, as this
function is marginally more appropriate than afterCheckRootProject.
Either function is fine (in global analysis mode, they are essentially
equivalent), but checkMergedProject is closer (in spirit) to what we
want, and afterCheckRootProject is usually called more often in partial
analysis mode (it is called for every module in the --analyze-only
phase), even though it will never do anything in
CommunicationDeviceDetector.

Bug: N/A
Test: existing
Change-Id: I053be760126b9c9eac03ccf60bd29adbee1ac448