Avoid initializing the JDK roots for Android modules
At some point, the UastEnvironment initialization started
unconditionally initializing JDK roots via a jar file
searches (LibraryUtils.findClassesFromJdkHome). On my
computer, this nearly doubled the time to run the lint
unit tests which made it frequently time out; I've worked
around using --test_timeout 1500 locally.

Most lint unit tests are Android tests; in Android, we
use the bootclasspath from android.jar, not from the JDK,
so this initialization is unnecessary. This CL updates
the driver to look for this condition and only initialize
when the JDK is needed. As a result running the
//tools/base/lint/libs/lint-tests:tests.test target
drops from 1430s to 986s locally.

Test: Existing
Bug: Not filed
Change-Id: I5ef6f1843868c5a42cb301b78aa81b09cde497e3