368059214: Support methods and fields in import alias test mode
This applies to both Java and Kotlin, and affects both the import alias
test mode and the fully qualified names test mode.

It also fixes a handful of detectors to correctly handle these, as
revealed by the new test modes.

Some of the initial failures weren't actually bugs in the detectors; it
was because the tests themselves were invalid; they had duplicate
classes. This typically happens when we have a Java test and then we
also add a test file for Kotlin and accidentally use the same class name
in Kotlin. When files are analyzed separately, this generally works --
but in the fully qualified test mode scenario, resolve sometimes picks
the other (duplicate) class. This isn't a valid source scenario; when
you build (and in the IDE) there's a duplicate class error. So now the
lint test infrastructure makes sure that across the test files in the
same package and source set, the main class names are unique.

Fixes: 368059214
Test: Many new and updated tests
Change-Id: I8aa37a021221fc74575001ac16663138168db2df